requirements.txt
flask
firebase
firebase_admin
python-dotenv